requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
plotly>=5.18.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pyathena>=3.0.0
sshtunnel>=0.4.0
